+++ src/com/fastorder/manager/IOrderManager.java
@@ -14,7 +14,7 @@ public interface IOrderManager {
 	public List<Product> getProducts(String[] productsId);
 	public Order getOrder(int id);
 	
-	public void create(List<Product> products, EstimatedTimeEnum estimatedTime, OrderStatusEnum status, float totalPrice, int userId, int shopId);
+	public void create(List<Product> products, EstimatedTimeEnum estimatedTime, OrderStatusEnum status, double totalPrice, int userId, int shopId);
 	public void updateOrderToInProgress(int idOrder);
 	
 	public void delete(int id);
